oath : corriger le debordement uint16_t d'oath_tlv_find, couvrir la garde hmac_len
Revue critique : (uint16_t)(i + 2u) <= len retronquait avant de comparer,
laissant une lecture d'un octet au-dela du tampon quand i approche 65534
avec len=65535. Corrige en comparant en 32 bits, comme la garde voisine.
Reproduit sous AddressSanitizer (heap-buffer-overflow) et couvert par un
test dedie (page memoire protegee + SIGSEGV catche, un simple sentinelle
d'octet ne pouvait pas faire mordre ce test : preuve par simulation
exhaustive que la garde interne masque toujours l'effet sur la valeur de
retour). Ajoute aussi la couverture manquante de la garde hmac_len < 20
d'oath_dynamic_binary, jamais exercee jusqu'ici.

+++ main/security/oath_proto.c
@@ -17,7 +17,12 @@ bool oath_tlv_find(const uint8_t *buf, uint16_t len, uint8_t tag,
 {
     if (buf == NULL) return false;
     uint16_t i = 0;
-    while ((uint16_t)(i + 2u) <= len) {
+    /* Comparaison en 32 bits, comme la garde voisine : un cast (uint16_t)
+     * ici retronquerait i+2 avant la comparaison, et quand i approche 65534
+     * avec len=65535, (uint16_t)(65534+2)=(uint16_t)65536=0 rendrait la
+     * garde vraie a tort — la boucle lirait alors un octet au-dela du
+     * tampon logique avant que la garde interne ne le rattrape. */
+    while ((uint32_t)i + 2u <= len) {
         const uint8_t t = buf[i];
         const uint8_t l = buf[i + 1];
         /* Longueur qui deborde : trame malformee, on s'arrete au lieu de lire
